[BugFix] Fix prune unused predicate column bug
